+++ app/src/main/AndroidManifest.xml
@@ -16,6 +16,12 @@
             android:name=".MainActivity"
             android:exported="true"
             android:label="@string/app_name"
+            android:theme="@style/Theme.NOWSOPTAndroid"/>
+
+        <activity
+            android:name=".LoginActivity"
+            android:exported="true"
+            android:label="@string/app_name"
             android:theme="@style/Theme.NOWSOPTAndroid">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -23,6 +29,14 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
+        <activity
+            android:name=".SignupActivity"
+            android:exported="true"
+            android:label="@string/app_name"
+            android:theme="@style/Theme.NOWSOPTAndroid"/>
+
+
     </application>
 
 </manifest>
